tests/wpt/tests/css/css-logical/parsing/padding-block-inline-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: parsing padding-block and padding-inline with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-padding-block">
<meta name="assert" content="padding-block, padding-inline support only the grammar '<'padding-top'>{1,2}'.">
<meta name="assert" content="padding-block, padding-inline longhands support only the grammar '<'padding-top'>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("padding-block-start", "none");
test_invalid_value("padding-block-start", "-10px");
test_invalid_value("padding-block-end", "1px 2px");
test_invalid_value("padding-block-end", "auto");
test_invalid_value("padding-block-end", "10");
test_invalid_value("padding-block-end", "1px, 2px");
test_invalid_value("padding-inline-start", "20% calc(10px - 0.5em)");
test_invalid_value("padding-inline-start", "2px auto");
test_invalid_value("padding-inline-end", "-20%");
test_invalid_value("padding-inline-end", "10px, auto");

test_invalid_value("padding-block", "none");
test_invalid_value("padding-block", "20% -10px");
test_invalid_value("padding-block", "auto, -10px");
test_invalid_value("padding-block", "20%, calc(10px - 0.5em)");
test_invalid_value("padding-inline", "10px auto 20px");
test_invalid_value("padding-inline", "1px 2px 3px");
test_invalid_value("padding-inline", "-20% calc(20% + 10px)");
</script>
</body>
</html>
